+++ source/extensions/isaacsim.robot_motion.cumotion.examples/isaacsim/robot_motion/cumotion/examples/trajectory_optimizer/scenario.py
@@ -31,7 +31,6 @@ if os.name != "nt":
 from isaacsim.robot_motion.cumotion.impl.utils import isaac_sim_to_cumotion_pose
 from isaacsim.robot_motion.experimental.motion_generation import (
     ObstacleConfiguration,
-    ObstacleRepresentation,
     ObstacleStrategy,
     SceneQuery,
     TrackableApi,
@@ -117,27 +116,9 @@ class FrankaTrajectoryOptimizerExample:
         print("Objects: ", objects)
 
         obstacle_strategy = ObstacleStrategy()
-        obstacle_strategy.set_default_configuration(
-            Mesh,
-            ObstacleConfiguration(
-                representation=ObstacleRepresentation.OBB,
-                safety_tolerance=0.01,
-            ),
-        )
-        obstacle_strategy.set_default_configuration(
-            Cone,
-            ObstacleConfiguration(
-                representation=ObstacleRepresentation.OBB,
-                safety_tolerance=0.01,
-            ),
-        )
-        obstacle_strategy.set_default_configuration(
-            Cylinder,
-            ObstacleConfiguration(
-                representation=ObstacleRepresentation.OBB,
-                safety_tolerance=0.01,
-            ),
-        )
+        obstacle_strategy.set_default_configuration(Mesh, ObstacleConfiguration("obb", 0.01))
+        obstacle_strategy.set_default_configuration(Cone, ObstacleConfiguration("obb", 0.01))
+        obstacle_strategy.set_default_configuration(Cylinder, ObstacleConfiguration("obb", 0.01))
 
         world_binding = WorldBinding(
             world_interface=CumotionWorldInterface(visualize_debug_prims=True),